fix: Rename "Resource" to "Res" because of PHP 7.0 soft reserved words and remove return type declaration to be compatible with PHP 5.6
