+++ inc/api/Res/Exception/AuthorizationException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace IgnicoWordPress\Api\Res\Exception;
+
+/**
+ * Throws when authorization failed and we know why e.g. we have message to show
+ * to user.
+ */
+class AuthorizationException extends \Exception {}
